kickbot/pyproject.toml
[project]
name = "kickbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Lorenzo van der Woude",email = "Lvanderwoude@hotmail.nl"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py (>=2.6.3,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]

[tool.poetry]
packages = [{include = "kickbot", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
